+++ static/js/mo/execute-command.js
@@ -1,0 +1,2 @@
+// execute_command.html
+$('#tab_server').addClass('active in')